requirements.txt
torch
torchaudio
torchvision
torchtext
torchmetrics
hydra-core
omegaconf
einops
flashy
sentencepiece
julius
numpy
scipy
soundfile
spacy
tqdm
transformers
xformers
num2words
librosa